Simplify generate.TP.fullgrids()
Several changes and refactorings to streamline this function.  Most
notably, some of the elements of the output were removed.  The inverse
transform functions were inputs to this function; thus, there was no
need to copy them to the output.  The coordinate array is already
stored in the griddata object, so there is also no need to recreate it
here.